Add permissions to installed binaries
Adds explicit ownership and permissions to installed binaries for a
number of programs installed on Linux using this playbook. This resolves
an issue with ansible-lint complaining about file permissions not being
mentioned. However, because the Ansible unarchive module does not
included the 'mode: preserve' option, a this rule will need to be added
to the ansible-lint configuration file skip list.