Ignore errors when installing Slack
Adds the ignore_errors flag to the play to install Slack, which returns
an error if there is a later version of the package installed than the
version downloaded by the play.

The should hopefully be only a temporary fix. Ideally, the play will be
skipped when there is a later version installed. However, like most
temporary fixes, I expect this will be very long-lasting.